style: refine the type to the file
