+++ picidae/date/detail/year.hpp
@@ -1,0 +1,14 @@
+#ifndef PICIDAE_DATE_DETAIL_YEAR_HPP__
+#define PICIDAE_DATE_DETAIL_YEAR_HPP__
+
+#include "picidae/meta/macro.hpp"
+
+PICIDAE_NAMESPACE_BEGIN(PICIDAE_NAMESPACE)
+
+PICIDAE_NAMESPACE_BEGIN(date)
+
+PICIDAE_NAMESPACE_END(date)
+
+PICIDAE_NAMESPACE_END(PICIDAE_NAMESPACE)
+
+#endif  // PICIDAE_DATE_DETAIL_YEAR_HPP__
